# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API Chat",
      "contact": "brunoleoncalderon18@gmail.com",
      "version": "1.0"
    },
    "paths": {
      "/api/v1/users": {
        "post": {
          "tags": [
            "users"
          ],
          "summary": "post register",
          "operationId": "postregister",
          "parameters": [],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/postregisterrequest"
                    },
                    {
                      "example": {
                        "firstname": "bruno",
                        "lastname": "leon",
                        "email": "brunoleoncalderon18@gmail.com",
                        "password": "Tatiana$28"
                      }
                    }
                  ]
                },
                "example": {
                  "firstname": "bruno",
                  "lastname": "leon",
                  "email": "brunoleoncalderon18@gmail.com",
                  "password": "Tatiana$28"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false
        },
        "get": {
          "tags": [
            "users"
          ],
          "summary": "get all Users",
          "operationId": "getallUsers",
          "parameters": [],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false,
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
...
      "/api/v1/messages/conversation/{id}": {
        "post": {
          "tags": [
            "messages"
          ],
          "summary": "Post messages",
          "operationId": "Postmessages",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32",
                "example": 5
              }
            }
          ],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PostmessagesRequest"
                    },
                    {
                      "example": {
                        "content": "que bueno ",
                        "senderId": 4
                      }
                    }
                  ]
                },
                "example": {
                  "content": "que bueno ",
                  "senderId": 4
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false,
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "get": {
          "tags": [
            "messages"
          ],
          "summary": "get Conversation Messages",
          "operationId": "getConversationMessages",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32",
                "example": 5
              }
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false,
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
...
    },
    "components": {
      "schemas": {
        "postregisterrequest": {
          "title": "postregisterrequest",
          "required": [
            "firstname",
            "lastname",
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "firstname": {
              "type": "string"
            },
            "lastname": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "example": {
            "firstname": "bruno",
            "lastname": "leon",
            "email": "brunoleoncalderon18@gmail.com",
            "password": "Tatiana$28"
          }
        },
...
        "PostmessagesRequest": {
          "title": "PostmessagesRequest",
          "required": [
            "content",
            "senderId"
          ],
          "type": "object",
          "properties": {
            "content": {
              "type": "string"
            },
            "senderId": {
              "type": "integer",
              "format": "int32"
            }
          },
          "example": {
            "content": "que bueno ",
            "senderId": 4
          }
        }
      },
      "securitySchemes": {
        "bearer": {
          "type": "http",
          "scheme": "bearer"
        }
      }
    },
    "security": [],
...
  }